% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transfer_decimal_number.R
\name{transfer_number}
\alias{transfer_number}
\title{Transfer decimal number}
\usage{
transfer_number(n, k = 2)
}
\description{
This function transfer decimal number to others
}
\examples{

transfer_number()
}
